bit_manipulation

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -1,3 +1,6 @@
+#include "main.h"
+#include <stdio.h>
+
 /**
  * flip_bits - function that returns the number of bits you would need to flip
  * to get from one number to another.
@@ -8,7 +11,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned long int target, num_bit = 0;
+	unsigned int target, num_bit = 0;
 
 	target = n ^ m;
 	for (; target; target >>= 1)
